System.Directory.Internal

Description

Internal modules are always subject to change from version to version. The contents of this module are also platform-dependent, hence what is shown in the Hackage documentation may differ from what is actually available on your system.

Documentation

andM :: Monad m => m Bool -> m Bool -> m Bool

copyHandleData

Arguments

:: Handle
Source handle

-> Handle
Destination handle

-> IO ()

Copy data from one handle to another until end of file.

dropSpecialDotDirs :: [OsPath] -> [OsPath]

emptyListT :: forall (m :: Type -> Type) a. Applicative m => ListT m a

emptyToCurDir :: OsPath -> OsPath

Convert empty paths to the current directory, otherwise leave it unchanged.

expandDots :: [OsPath] -> [OsPath]

Given a list of path segments, expand . and ... The path segments must not contain path separators.

fileTypeIsDirectory :: FileType -> Bool

Check whether the given FileType is considered a directory by the operating system. This affects the choice of certain functions e.g. removeDirectory vs removeFile.

fileTypeIsLink :: FileType -> Bool

Return whether the given FileType is a link.

ignoreIOExceptions :: IO () -> IO ()

Attempt to perform the given action, silencing any IO exception thrown by it.

ioeAddLocation :: IOError -> String -> IOError

ioeSetOsPath :: IOError -> OsPath -> IOError

isNoFollow :: WhetherFollow -> Bool

liftJoinListT :: Monad m => m (ListT m a) -> ListT m a

listTHead :: Functor m => ListT m a -> m (Maybe a)

listTToList :: Monad m => ListT m a -> m [a]

listToListT :: forall (m :: Type -> Type) a. Applicative m => [a] -> ListT m a

maybeToListT :: Applicative m => m (Maybe a) -> ListT m a

normalisePathSeps :: OsPath -> OsPath

Convert to the right kind of slashes.

normaliseTrailingSep :: OsPath -> OsPath

Remove redundant trailing slashes and pick the right kind of slash.

os :: String -> OsString

Fallibly converts String to OsString. Only intended to be used on literals.

rightOrError :: Exception e => Either e a -> a

sequenceWithIOErrors_ :: [IO ()] -> IO ()

simplifyPosix :: OsPath -> OsPath

Similar to normalise but empty paths stay empty.

simplifyWindows :: OsPath -> OsPath

Similar to normalise but:

empty paths stay empty,

parent dirs (..) are expanded, and

paths starting with \\?\ are preserved.

The goal is to preserve the meaning of paths better than normalise.

so :: OsString -> String

Fallibly converts OsString to String. Only intended to be used on literals.

specializeErrorString :: String -> (IOError -> Bool) -> IO a -> IO a

tryIOErrorType :: (IOError -> Bool) -> IO a -> IO (Either IOError a)

Similar to try but only catches a specify kind of IOError as specified by the predicate.

withBinaryFile :: OsPath -> IOMode -> (Handle -> IO r) -> IO r

data FileType

Constructors

File

SymbolicLink
POSIX: either file or directory link; Windows: file link

Directory

DirectoryLink
Windows only: directory link

newtype ListT (m :: Type -> Type) a

A generator with side-effects.

Constructors

ListT

Fields

unListT :: m (Maybe (a, ListT m a))

data Permissions

Constructors

Permissions

Fields

readable :: Bool

writable :: Bool

executable :: Bool

searchable :: Bool

data WhetherFollow

Constructors

NoFollow

FollowLinks

data XdgDirectory

Special directories for storing user-specific application data, configuration, and cache files, as specified by the XDG Base Directory Specification.

Note: On Windows, XdgData and XdgConfig usually map to the same directory.

Since: directory-1.2.3.0

Constructors

XdgData
For data files (e.g. images). It uses the XDG_DATA_HOME environment variable. On non-Windows systems, the default is ~/.local/share. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /usr/share.

XdgConfig
For configuration files. It uses the XDG_CONFIG_HOME environment variable. On non-Windows systems, the default is ~/.config. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /etc.

XdgCache
For non-essential files (e.g. cache). It uses the XDG_CACHE_HOME environment variable. On non-Windows systems, the default is ~/.cache. On Windows, the default is %LOCALAPPDATA% (e.g. C:/Users/<user>/AppData/Local). Can be considered as the user-specific equivalent of /var/cache.

XdgState
For data that should persist between (application) restarts, but that is not important or portable enough to the user that it should be stored in XdgData. It uses the XDG_STATE_HOME environment variable. On non-Windows sytems, the default is ~/.local/state. On Windows, the default is %LOCALAPPDATA% (e.g. C:/Users/<user>/AppData/Local).

Since: directory-1.3.7.0

data XdgDirectoryList

Search paths for various application data, as specified by the XDG Base Directory Specification.

The list of paths is split using searchPathSeparator, which on Windows is a semicolon.

Note: On Windows, XdgDataDirs and XdgConfigDirs usually yield the same result.

Since: directory-1.3.2.0

Constructors

XdgDataDirs
For data files (e.g. images). It uses the XDG_DATA_DIRS environment variable. On non-Windows systems, the default is /usr/local/share/ and /usr/share/. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

XdgConfigDirs
For configuration files. It uses the XDG_CONFIG_DIRS environment variable. On non-Windows systems, the default is /etc/xdg. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

data OsString

Newtype representing short operating system specific strings.

Internally this is either WindowsString or PosixString, depending on the platform. Both use unpinned ShortByteString for efficiency.

The constructor is only exported via System.OsString.Internal.Types, since dealing with the internals isn't generally recommended, but supported in case you need to write platform specific code.

accessTimeFromMetadata :: Metadata -> UTCTime

allWriteMode :: FileMode

atWhetherFollow :: WhetherFollow -> CInt

c_AT_FDCWD :: Fd

c_AT_SYMLINK_NOFOLLOW :: CInt

c_PATH_MAX :: Maybe Int

c_fchmodat :: Fd -> CString -> FileMode -> CInt -> IO CInt

c_free :: Ptr a -> IO ()

c_fstatat :: Fd -> CString -> Ptr CStat -> CInt -> IO CInt

c_realpath :: CString -> CString -> IO CString

c_unlinkat :: Fd -> CString -> CInt -> IO CInt

canonicalizePathSimplify :: OsPath -> IO OsPath

closeRaw :: RawHandle -> IO ()

copyFileContents

Arguments

:: OsPath
Source filename

-> OsPath
Destination filename

-> IO ()

Truncate the destination file and then copy the contents of the source file to the destination file. If the destination file already exists, its attributes shall remain unchanged. Otherwise, its attributes are reset to the defaults.

copyFileWithMetadataInternal :: (Metadata -> OsPath -> IO ()) -> (Metadata -> OsPath -> IO ()) -> OsPath -> OsPath -> IO ()

copyGroupFromStatus :: FileStatus -> OsPath -> IO ()

copyOwnerFromStatus :: FileStatus -> OsPath -> IO ()

createDirectoryInternal :: OsPath -> IO ()

createHardLink :: OsPath -> OsPath -> IO ()

createSymbolicLink :: Bool -> OsPath -> OsPath -> IO ()

defaultOpenFlags :: OpenFileFlags

exeExtensionInternal :: OsString

fileSizeFromMetadata :: Metadata -> Integer

fileTypeFromMetadata :: Metadata -> FileType

filesAlwaysRemovable :: Bool

findExecutablesLazyInternal :: ([OsPath] -> OsString -> ListT IO OsPath) -> OsString -> ListT IO OsPath

getAccessPermissions :: OsPath -> IO Permissions

getAppUserDataDirectoryInternal :: OsPath -> IO OsPath

getCurrentDirectoryInternal :: IO OsPath

getDirectoryContentsInternal :: OsPath -> IO [OsPath]

getEnvOs :: OsString -> IO OsString

getFileMetadata :: OsPath -> IO Metadata

getHomeDirectoryInternal :: IO OsPath

$HOME is preferred, because the user has control over it. However, POSIX doesn't define it as a mandatory variable, so fall back to getpwuid_r.

getMetadataAt :: WhetherFollow -> Maybe RawHandle -> OsPath -> IO Metadata

getPath :: IO [OsPath]

Get the contents of the PATH environment variable.

getSymbolicLinkMetadata :: OsPath -> IO Metadata

getTemporaryDirectoryInternal :: IO OsPath

getUserDocumentsDirectoryInternal :: IO OsPath

getXdgDirectoryFallback :: IO OsPath -> XdgDirectory -> IO OsPath

getXdgDirectoryListFallback :: XdgDirectoryList -> IO [OsPath]

hasWriteMode :: Mode -> Bool

linkToDirectoryIsDirectory :: Bool

lookupEnvOs :: OsString -> IO (Maybe OsString)

modeFromMetadata :: Metadata -> Mode

modificationTimeFromMetadata :: Metadata -> UTCTime

openDirFromFd :: Fd -> IO DirStream

openRaw :: WhetherFollow -> Maybe RawHandle -> OsPath -> IO RawHandle

prependCurrentDirectory :: OsPath -> IO OsPath

Convert a path into an absolute path. If the given path is relative, the current directory is prepended and the path may or may not be simplified. If the path is already absolute, the path is returned unchanged. The function preserves the presence or absence of the trailing path separator.

If the path is already absolute, the operation never fails. Otherwise, the operation may throw exceptions.

Empty paths are treated as the current directory.

readDirStreamToEnd :: DirStream -> IO [OsPath]

readDirToEnd :: RawHandle -> IO [OsPath]

readSymbolicLink :: OsPath -> IO OsPath

realPath :: OsPath -> IO OsPath

removePathAt :: FileType -> Maybe RawHandle -> OsPath -> IO ()

removePathInternal :: Bool -> OsPath -> IO ()

renamePathInternal :: OsPath -> OsPath -> IO ()

setAccessPermissions :: OsPath -> Permissions -> IO ()

setCurrentDirectoryInternal :: OsPath -> IO ()

setFileMode :: OsPath -> Mode -> IO ()

setFilePermissions :: OsPath -> Mode -> IO ()

setForceRemoveMode :: Mode -> Mode

setModeAt :: WhetherFollow -> Maybe RawHandle -> OsPath -> Mode -> IO ()

setTimes :: OsPath -> (Maybe POSIXTime, Maybe POSIXTime) -> IO ()

setWriteMode :: Bool -> Mode -> Mode

simplify :: OsPath -> OsPath

On POSIX, equivalent to simplifyPosix.

tryCopyOwnerAndGroupFromStatus :: FileStatus -> OsPath -> IO ()

withRealpath :: CString -> (CString -> IO a) -> IO a

type Metadata = FileStatus

type Mode = FileMode

type RawHandle = Fd