System.Directory.Internal

Description

Internal modules are always subject to change from version to version. The contents of this module are also platform-dependent, hence what is shown in the Hackage documentation may differ from what is actually available on your system.

Documentation

data XdgDirectory

Special directories for storing user-specific application data, configuration, and cache files, as specified by the XDG Base Directory Specification.

Note: On Windows, XdgData and XdgConfig usually map to the same directory.

Since: directory-1.2.3.0

Constructors

XdgData
For data files (e.g. images). It uses the XDG_DATA_HOME environment variable. On non-Windows systems, the default is ~/.local/share. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /usr/share.

XdgConfig
For configuration files. It uses the XDG_CONFIG_HOME environment variable. On non-Windows systems, the default is ~/.config. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /etc.

XdgCache
For non-essential files (e.g. cache). It uses the XDG_CACHE_HOME environment variable. On non-Windows systems, the default is ~/.cache. On Windows, the default is %LOCALAPPDATA% (e.g. C:/Users/<user>/AppData/Local). Can be considered as the user-specific equivalent of /var/cache.

XdgState
For data that should persist between (application) restarts, but that is not important or portable enough to the user that it should be stored in XdgData. It uses the XDG_STATE_HOME environment variable. On non-Windows sytems, the default is ~/.local/state. On Windows, the default is %LOCALAPPDATA% (e.g. C:/Users/<user>/AppData/Local).

Since: directory-1.3.7.0

data XdgDirectoryList

Search paths for various application data, as specified by the XDG Base Directory Specification.

The list of paths is split using searchPathSeparator, which on Windows is a semicolon.

Note: On Windows, XdgDataDirs and XdgConfigDirs usually yield the same result.

Since: directory-1.3.2.0

Constructors

XdgDataDirs
For data files (e.g. images). It uses the XDG_DATA_DIRS environment variable. On non-Windows systems, the default is /usr/local/share/ and /usr/share/. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

XdgConfigDirs
For configuration files. It uses the XDG_CONFIG_DIRS environment variable. On non-Windows systems, the default is /etc/xdg. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

data Permissions

Constructors

Permissions

Fields

readable :: Bool

writable :: Bool

executable :: Bool

searchable :: Bool

newtype ListT m a

A generator with side-effects.

Constructors

ListT

Fields

unListT :: m (Maybe (a, ListT m a))

data FileType

Constructors

File

SymbolicLink
POSIX: either file or directory link; Windows: file link

Directory

DirectoryLink
Windows only: directory link

os :: String -> OsString

Fallibly converts String to OsString. Only intended to be used on literals.

emptyListT :: Applicative m => ListT m a

maybeToListT :: Applicative m => m (Maybe a) -> ListT m a

listToListT :: Applicative m => [a] -> ListT m a

liftJoinListT :: Monad m => m (ListT m a) -> ListT m a

listTHead :: Functor m => ListT m a -> m (Maybe a)

listTToList :: Monad m => ListT m a -> m [a]

andM :: Monad m => m Bool -> m Bool -> m Bool

sequenceWithIOErrors_ :: [IO ()] -> IO ()

tryIOErrorType :: (IOError -> Bool) -> IO a -> IO (Either IOError a)

Similar to try but only catches a specify kind of IOError as specified by the predicate.

ignoreIOExceptions :: IO () -> IO ()

Attempt to perform the given action, silencing any IO exception thrown by it.

specializeErrorString :: String -> (IOError -> Bool) -> IO a -> IO a

ioeAddLocation :: IOError -> String -> IOError

rightOrError :: Exception e => Either e a -> a

so :: OsString -> String

Fallibly converts OsString to String. Only intended to be used on literals.

ioeSetOsPath :: IOError -> OsPath -> IOError

expandDots :: [OsPath] -> [OsPath]

Given a list of path segments, expand . and ... The path segments must not contain path separators.

normalisePathSeps :: OsPath -> OsPath

Convert to the right kind of slashes.

normaliseTrailingSep :: OsPath -> OsPath

Remove redundant trailing slashes and pick the right kind of slash.

emptyToCurDir :: OsPath -> OsPath

Convert empty paths to the current directory, otherwise leave it unchanged.

simplifyPosix :: OsPath -> OsPath

Similar to normalise but empty paths stay empty.

simplifyWindows :: OsPath -> OsPath

Similar to normalise but:

empty paths stay empty,

parent dirs (..) are expanded, and

paths starting with \\?\ are preserved.

The goal is to preserve the meaning of paths better than normalise.

fileTypeIsDirectory :: FileType -> Bool

Check whether the given FileType is considered a directory by the operating system. This affects the choice of certain functions e.g. removeDirectory vs removeFile.

fileTypeIsLink :: FileType -> Bool

Return whether the given FileType is a link.

withBinaryHandle :: IO Handle -> (Handle -> IO r) -> IO r

copyHandleData

Arguments

:: Handle
Source handle

-> Handle
Destination handle

-> IO ()

Copy data from one handle to another until end of file.

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

data OsString

Newtype representing short operating system specific strings.

Internally this is either WindowsString or PosixString, depending on the platform. Both use unpinned ShortByteString for efficiency.

The constructor is only exported via System.OsString.Internal.Types, since dealing with the internals isn't generally recommended, but supported in case you need to write platform specific code.

createDirectoryInternal :: OsPath -> IO ()

removePathInternal :: Bool -> OsPath -> IO ()

renamePathInternal :: OsPath -> OsPath -> IO ()

filesAlwaysRemovable :: Bool

simplify :: OsPath -> OsPath

On POSIX, equivalent to simplifyPosix.

c_free :: Ptr a -> IO ()

c_PATH_MAX :: Maybe Int

c_realpath :: CString -> CString -> IO CString

withRealpath :: CString -> (CString -> IO a) -> IO a

realPath :: OsPath -> IO OsPath

canonicalizePathSimplify :: OsPath -> IO OsPath

findExecutablesLazyInternal :: ([OsPath] -> OsString -> ListT IO OsPath) -> OsString -> ListT IO OsPath

exeExtensionInternal :: OsString

getDirectoryContentsInternal :: OsPath -> IO [OsPath]

getCurrentDirectoryInternal :: IO OsPath

prependCurrentDirectory :: OsPath -> IO OsPath

Convert a path into an absolute path. If the given path is relative, the current directory is prepended and the path may or may not be simplified. If the path is already absolute, the path is returned unchanged. The function preserves the presence or absence of the trailing path separator.

If the path is already absolute, the operation never fails. Otherwise, the operation may throw exceptions.

Empty paths are treated as the current directory.

setCurrentDirectoryInternal :: OsPath -> IO ()

linkToDirectoryIsDirectory :: Bool

createHardLink :: OsPath -> OsPath -> IO ()

createSymbolicLink :: Bool -> OsPath -> OsPath -> IO ()

readSymbolicLink :: OsPath -> IO OsPath

type Metadata = FileStatus

getSymbolicLinkMetadata :: OsPath -> IO Metadata

getFileMetadata :: OsPath -> IO Metadata

fileTypeFromMetadata :: Metadata -> FileType

fileSizeFromMetadata :: Metadata -> Integer

accessTimeFromMetadata :: Metadata -> UTCTime

modificationTimeFromMetadata :: Metadata -> UTCTime

type Mode = FileMode

modeFromMetadata :: Metadata -> Mode

allWriteMode :: FileMode

hasWriteMode :: Mode -> Bool

setWriteMode :: Bool -> Mode -> Mode

setFileMode :: OsPath -> Mode -> IO ()

setFilePermissions :: OsPath -> Mode -> IO ()

getAccessPermissions :: OsPath -> IO Permissions

setAccessPermissions :: OsPath -> Permissions -> IO ()

copyOwnerFromStatus :: FileStatus -> OsPath -> IO ()

copyGroupFromStatus :: FileStatus -> OsPath -> IO ()

tryCopyOwnerAndGroupFromStatus :: FileStatus -> OsPath -> IO ()

defaultFlags :: OpenFileFlags

openFileForRead :: OsPath -> IO Handle

openFileForWrite :: OsPath -> IO Handle

copyFileContents

Arguments

:: OsPath
Source filename

-> OsPath
Destination filename

-> IO ()

Truncate the destination file and then copy the contents of the source file to the destination file. If the destination file already exists, its attributes shall remain unchanged. Otherwise, its attributes are reset to the defaults.

copyFileWithMetadataInternal :: (Metadata -> OsPath -> IO ()) -> (Metadata -> OsPath -> IO ()) -> OsPath -> OsPath -> IO ()

setTimes :: OsPath -> (Maybe POSIXTime, Maybe POSIXTime) -> IO ()

lookupEnvOs :: OsString -> IO (Maybe OsString)

getEnvOs :: OsString -> IO OsString

getPath :: IO [OsPath]

Get the contents of the PATH environment variable.

getHomeDirectoryInternal :: IO OsPath

$HOME is preferred, because the user has control over it. However, POSIX doesn't define it as a mandatory variable, so fall back to getpwuid_r.

getXdgDirectoryFallback :: IO OsPath -> XdgDirectory -> IO OsPath

getXdgDirectoryListFallback :: XdgDirectoryList -> IO [OsPath]

getAppUserDataDirectoryInternal :: OsPath -> IO OsPath

getUserDocumentsDirectoryInternal :: IO OsPath

getTemporaryDirectoryInternal :: IO OsPath