System.Directory.Internal

Description

Internal modules are always subject to change from version to version. The contents of this module are also platform-dependent, hence what is shown in the Hackage documentation may differ from what is actually available on your system.

Documentation

newtype ListT m a

A generator with side-effects.

Constructors

ListT

Fields

unListT :: m (Maybe (a, ListT m a))

emptyListT :: Applicative m => ListT m a

maybeToListT :: Applicative m => m (Maybe a) -> ListT m a

listToListT :: Applicative m => [a] -> ListT m a

liftJoinListT :: Monad m => m (ListT m a) -> ListT m a

listTHead :: Functor m => ListT m a -> m (Maybe a)

listTToList :: Monad m => ListT m a -> m [a]

andM :: Monad m => m Bool -> m Bool -> m Bool

sequenceWithIOErrors_ :: [IO ()] -> IO ()

tryIOErrorType :: (IOError -> Bool) -> IO a -> IO (Either IOError a)

Similar to try but only catches a specify kind of IOError as specified by the predicate.

ignoreIOExceptions :: IO () -> IO ()

Attempt to perform the given action, silencing any IO exception thrown by it.

specializeErrorString :: String -> (IOError -> Bool) -> IO a -> IO a

ioeAddLocation :: IOError -> String -> IOError

expandDots :: [FilePath] -> [FilePath]

Given a list of path segments, expand . and ... The path segments must not contain path separators.

normalisePathSeps :: FilePath -> FilePath

Convert to the right kind of slashes.

normaliseTrailingSep :: FilePath -> FilePath

Remove redundant trailing slashes and pick the right kind of slash.

emptyToCurDir :: FilePath -> FilePath

Convert empty paths to the current directory, otherwise leave it unchanged.

simplifyPosix :: FilePath -> FilePath

Similar to normalise but empty paths stay empty.

simplifyWindows :: FilePath -> FilePath

Similar to normalise but:

empty paths stay empty,

parent dirs (..) are expanded, and

paths starting with \\?\ are preserved.

The goal is to preserve the meaning of paths better than normalise.

data FileType

Constructors

File

SymbolicLink
POSIX: either file or directory link; Windows: file link

Directory

DirectoryLink
Windows only: directory link

fileTypeIsDirectory :: FileType -> Bool

Check whether the given FileType is considered a directory by the operating system. This affects the choice of certain functions e.g. removeDirectory vs removeFile.

fileTypeIsLink :: FileType -> Bool

Return whether the given FileType is a link.

data Permissions

Constructors

Permissions

Fields

readable :: Bool

writable :: Bool

executable :: Bool

searchable :: Bool

copyFileContents

Arguments

:: FilePath
Source filename

-> FilePath
Destination filename

-> IO ()

Truncate the destination file and then copy the contents of the source file to the destination file. If the destination file already exists, its attributes shall remain unchanged. Otherwise, its attributes are reset to the defaults.

copyFileToHandle

Arguments

:: FilePath
Source file

-> Handle
Destination handle

-> IO ()

Copy all data from a file to a handle.

copyHandleData

Arguments

:: Handle
Source handle

-> Handle
Destination handle

-> IO ()

Copy data from one handle to another until end of file.

data XdgDirectory

Special directories for storing user-specific application data, configuration, and cache files, as specified by the XDG Base Directory Specification.

Note: On Windows, XdgData and XdgConfig usually map to the same directory.

Since: directory-1.2.3.0

Constructors

XdgData
For data files (e.g. images). It uses the XDG_DATA_HOME environment variable. On non-Windows systems, the default is ~/.local/share. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /usr/share.

XdgConfig
For configuration files. It uses the XDG_CONFIG_HOME environment variable. On non-Windows systems, the default is ~/.config. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /etc.

XdgCache
For non-essential files (e.g. cache). It uses the XDG_CACHE_HOME environment variable. On non-Windows systems, the default is ~/.cache. On Windows, the default is %LOCALAPPDATA% (e.g. C:/Users/<user>/AppData/Local). Can be considered as the user-specific equivalent of /var/cache.

data XdgDirectoryList

Search paths for various application data, as specified by the XDG Base Directory Specification.

The list of paths is split using searchPathSeparator, which on Windows is a semicolon.

Note: On Windows, XdgDataDirs and XdgConfigDirs usually yield the same result.

Since: directory-1.3.2.0

Constructors

XdgDataDirs
For data files (e.g. images). It uses the XDG_DATA_DIRS environment variable. On non-Windows systems, the default is /usr/local/share/ and /usr/share/. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

XdgConfigDirs
For configuration files. It uses the XDG_CONFIG_DIRS environment variable. On non-Windows systems, the default is /etc/xdg. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

createDirectoryInternal :: FilePath -> IO ()

removePathInternal :: Bool -> FilePath -> IO ()

renamePathInternal :: FilePath -> FilePath -> IO ()

simplify :: FilePath -> FilePath

On POSIX, equivalent to simplifyPosix.

c_free :: Ptr a -> IO ()

c_PATH_MAX :: Maybe Int

c_realpath :: CString -> CString -> IO CString

withRealpath :: CString -> (CString -> IO a) -> IO a

canonicalizePathWith :: ((FilePath -> IO FilePath) -> FilePath -> IO FilePath) -> FilePath -> IO FilePath

canonicalizePathSimplify :: FilePath -> IO FilePath

findExecutablesLazyInternal :: ([FilePath] -> String -> ListT IO FilePath) -> String -> ListT IO FilePath

exeExtensionInternal :: String

getDirectoryContentsInternal :: FilePath -> IO [FilePath]

getCurrentDirectoryInternal :: IO FilePath

prependCurrentDirectory :: FilePath -> IO FilePath

Convert a path into an absolute path. If the given path is relative, the current directory is prepended and the path may or may not be simplified. If the path is already absolute, the path is returned unchanged. The function preserves the presence or absence of the trailing path separator.

If the path is already absolute, the operation never fails. Otherwise, the operation may throw exceptions.

Empty paths are treated as the current directory.

setCurrentDirectoryInternal :: FilePath -> IO ()

linkToDirectoryIsDirectory :: Bool

createSymbolicLink :: Bool -> FilePath -> FilePath -> IO ()

readSymbolicLink :: FilePath -> IO FilePath

type Metadata = FileStatus

getSymbolicLinkMetadata :: FilePath -> IO Metadata

getFileMetadata :: FilePath -> IO Metadata

fileTypeFromMetadata :: Metadata -> FileType

fileSizeFromMetadata :: Metadata -> Integer

accessTimeFromMetadata :: Metadata -> UTCTime

modificationTimeFromMetadata :: Metadata -> UTCTime

posix_accessTimeHiRes :: FileStatus -> POSIXTime

posix_modificationTimeHiRes :: FileStatus -> POSIXTime

type Mode = FileMode

modeFromMetadata :: Metadata -> Mode

allWriteMode :: FileMode

hasWriteMode :: Mode -> Bool

setWriteMode :: Bool -> Mode -> Mode

setFileMode :: FilePath -> Mode -> IO ()

setFilePermissions :: FilePath -> Mode -> IO ()

getAccessPermissions :: FilePath -> IO Permissions

setAccessPermissions :: FilePath -> Permissions -> IO ()

copyOwnerFromStatus :: FileStatus -> FilePath -> IO ()

copyGroupFromStatus :: FileStatus -> FilePath -> IO ()

tryCopyOwnerAndGroupFromStatus :: FileStatus -> FilePath -> IO ()

copyFileWithMetadataInternal :: (Metadata -> FilePath -> IO ()) -> (Metadata -> FilePath -> IO ()) -> FilePath -> FilePath -> IO ()

setTimes :: FilePath -> (Maybe POSIXTime, Maybe POSIXTime) -> IO ()

getPath :: IO [FilePath]

Get the contents of the PATH environment variable.

getHomeDirectoryInternal :: IO FilePath

$HOME is preferred, because the user has control over it. However, POSIX doesn't define it as a mandatory variable, so fall back to getpwuid_r.

getXdgDirectoryFallback :: IO FilePath -> XdgDirectory -> IO FilePath

getXdgDirectoryListFallback :: XdgDirectoryList -> IO [FilePath]

getAppUserDataDirectoryInternal :: FilePath -> IO FilePath

getUserDocumentsDirectoryInternal :: IO FilePath

getTemporaryDirectoryInternal :: IO FilePath